Support SQLite by only doing full-text if using mysql

+++ db/migrate/009_full_text_search.rb
@@ -1,11 +1,15 @@
 class FullTextSearch < ActiveRecord::Migration
   def self.up
-    execute 'ALTER TABLE torrents ENGINE = MyISAM'
-    execute 'CREATE FULLTEXT INDEX ft_idx_torrents ON torrents(name,filename)'
+    if ActiveRecord::Base.configurations[ENV['RAILS_ENV']]["adapter"] == "mysql"
+      execute 'ALTER TABLE torrents ENGINE = MyISAM'
+      execute 'CREATE FULLTEXT INDEX ft_idx_torrents ON torrents(name,filename)'
+    end
   end
 
   def self.down
-    execute 'ALTER TABLE torrents DROP INDEX ft_idx_torrents'
-    execute 'ALTER TABLE torrents ENGINE = InnoDB'
+    if ActiveRecord::Base.configurations[ENV['RAILS_ENV']]["adapter"] == "mysql"
+      execute 'ALTER TABLE torrents DROP INDEX ft_idx_torrents'
+      execute 'ALTER TABLE torrents ENGINE = InnoDB'
+    end
   end
 end
