+++ db/migrate/20081122190700_add_full_text_to_torrents.rb
@@ -1,8 +1,10 @@
 class AddFullTextToTorrents < ActiveRecord::Migration
   def self.up
-    remove_index "torrents", :name => "ft_idx_torrents"
-    execute("ALTER TABLE torrents ENGINE = MyISAM")
-    execute("CREATE FULLTEXT INDEX FullText_torrents ON torrents (`name`, `filename`, `description`);")
+    if ActiveRecord::Base.configurations[RAILS_ENV]["adapter"] == "mysql"
+      remove_index "torrents", :name => "ft_idx_torrents"
+      execute("ALTER TABLE torrents ENGINE = MyISAM")
+      execute("CREATE FULLTEXT INDEX FullText_torrents ON torrents (`name`, `filename`, `description`);")
+    end
   end
 
   def self.down
